refactor: rename yoyConfig
